переполнение разрядности

+++ Syntez/si5351a.h
@@ -14,7 +14,7 @@
 class Si5351 {
   private:
     uint32_t xtal_freq, freq0, freq1, freq2, freq_pll_b;
-    uint8_t freq0_div, freq1_div, freq2_div;
+    uint16_t freq0_div, freq1_div, freq2_div;
     uint8_t freq0_rdiv, freq1_rdiv, freq2_rdiv;
     uint8_t power0, power1, power2;
     void update_freq0(uint8_t* need_reset_pll);
